requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cors==1.0.1
dj-config-url==0.1.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.10.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.4.0
zope.event==4.6
zope.interface==6.0
